backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
opencv-python==4.10.0.84
numpy==2.2.2
scipy==1.15.1
ffmpeg-python==0.2.0
python-multipart==0.0.20
google-generativeai==0.8.3
google-genai>=0.3.0
python-dotenv==1.0.1
Pillow==11.2.1
fal-client==0.11.0
requests==2.32.3
